pyproject.toml
[tool.poetry]
name = "s3zipda"
version = "0.1.0"
description = "Functionality to access data stored inside zip files, either locally or in S3"
authors = ["Eetu Mäkelä <eetu.makela@helsinki.fi>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
s3fs = "^2024.9.0"
morefs = {extras = ["asynclocalfs"], version = "^0.2.2"}
pyyaml = "^6.0.2"
hscifsspecutil = "^0.1.1"
stream-unzip = "^0.0.95"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
hereutil = "^0.1.5"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
